requirements.txt
pandas==1.3.0
Pillow==8.4.0
python-dotenv==0.19.1
selenium==3.141.0
zxing==0.13.1
